+++ test/bigtest/network/serializers/file-extension.js
@@ -2,7 +2,15 @@ import { RestSerializer } from '@bigtest/mirage';
 
 export default RestSerializer.extend({
   serialize(...args) {
-    const { fileExtensions } = RestSerializer.prototype.serialize.apply(this, args);
+    const json = RestSerializer.prototype.serialize.apply(this, args);
+    const {
+      fileExtension,
+      fileExtensions,
+    } = json;
+
+    if (fileExtension) {
+      return fileExtension;
+    }
 
     return {
       fileExtensions,